Allow running rg from Dired on root directory
Previously if you opened a Dired buffer to the root directory of your
filesystem and attempted to use rg, rg-project-root would get a nil
value for FILE. All of the backends would fail to provide a project
root, so rg-project-root would fall back on asking for the directory
that contains FILE. Unfortunately, in this case FILE is unset. The
problem is solved by handling the cases where FILE is nil (or doesn't
have a directory component), and defaulting to default-directory.